split ConsoleProcess APIs into separate source
